:recycle: restructured Streamer application and added input dialog to select the device
