+++ application/streamer.py
@@ -1,7 +1,4 @@
-import re
 import time
-import bluetooth
-import argparse
 import matplotlib
 import numpy as np
 import logging
@@ -9,9 +6,7 @@ import pyqtgraph as pg
 import os
 import pylsl
 from PyQt5 import QtWidgets, QtCore
-from PyQt5.QtCore import QThread, pyqtSignal
-from pylsl import resolve_stream, StreamInlet
-
+from application.lsl_stream import LSLStreamThread
 from utils.layouts import layouts
 from pyqtgraph import ScatterPlotItem, mkBrush
 from brainflow.board_shim import BoardShim, BrainFlowInputParams, BoardIds
@@ -24,36 +19,10 @@ colors = ["blue", "green", "yellow", "purple", "orange", "pink", "brown", "gray"
           "cyan", "magenta", "lime", "teal", "lavender", "turquoise", "maroon", "olive"]
 
 
-def retrieve_unicorn_devices():
-    saved_devices = bluetooth.discover_devices(duration=1, lookup_names=True, lookup_class=True)
-    unicorn_devices = list(filter(lambda x: re.search(r'UN-\d{4}.\d{2}.\d{2}', x[1]), saved_devices))
-    logging.info(f"Found Unicorns: {unicorn_devices} ")
-    return unicorn_devices
-
-
 def write_header(file, board_id):
     for column in layouts[board_id]["header"]:
         file.write(str(column) + '\t')
     file.write('\n')
-
-
-class LSLStreamThread(QThread):
-    new_sample = pyqtSignal(object, float)  # Signal to emit new sample data
-
-    def run(self):
-        # Resolve an LSL stream named 'MyStream'
-        logging.info("Looking for an LSL stream...")
-        streams = resolve_stream('type', 'Markers')
-
-        # Create a new inlet to read from the stream
-        inlet = StreamInlet(streams[0])
-
-        while True:
-            # Pull a new sample from the inlet
-            marker, timestamp = inlet.pull_sample()
-            logging.info({"got %s at time %s" % (marker[0], timestamp)})
-            # Emit the new sample data
-            self.new_sample.emit(marker[0], timestamp)
 
 
 class Streamer:
@@ -326,7 +295,6 @@ class Streamer:
             timestamp = time.time()
         self.board.insert_marker(int(trigger))
 
-
     def update_quality_indicators(self, sample):
         eeg_start = layouts[self.board_id]["eeg_start"]
         eeg_end = layouts[self.board_id]["eeg_end"]
@@ -361,69 +329,3 @@ class Streamer:
             self.board.start_stream()
             self.start_button.setText('Stop')
             self.is_streaming = True
-
-
-def main():
-    BoardShim.enable_dev_board_logger()
-    logging.basicConfig(level=logging.INFO)
-
-    parser = argparse.ArgumentParser()
-    # use docs to check which parameters are required for specific board, e.g. for Cyton - set serial port
-    parser.add_argument('--timeout', type=int, help='timeout for device discovery or connection', required=False,
-                        default=0)
-    parser.add_argument('--ip-port', type=int, help='ip port', required=False, default=0)
-    parser.add_argument('--ip-protocol', type=int, help='ip protocol, check IpProtocolType enum', required=False,
-                        default=0)
-    parser.add_argument('--ip-address', type=str, help='ip address', required=False, default='')
-    parser.add_argument('--serial-port', type=str, help='serial port', required=False, default='')
-    parser.add_argument('--mac-address', type=str, help='mac address', required=False, default='')
-    parser.add_argument('--other-info', type=str, help='other info', required=False, default='')
-    parser.add_argument('--streamer-params', type=str, help='streamer params', required=False, default='')
-    parser.add_argument('--serial-number', type=str, help='serial number', required=False, default='')
-    parser.add_argument('--board-id', type=int, help='board id, check docs to get a list of supported boards',
-                        required=False, default=BoardIds.SYNTHETIC_BOARD)
-    parser.add_argument('--file', type=str, help='file', required=False, default='')
-    parser.add_argument('--master-board', type=int, help='master board id for streaming and playback boards',
-                        required=False, default=BoardIds.NO_BOARD)
-    parser.add_argument('--unicorn-index', type=int, help='index of the unicorn device', required=False, default=0)
-    args = parser.parse_args()
-
-    params = BrainFlowInputParams()
-    params.ip_port = args.ip_port
-    params.serial_port = args.serial_port
-    params.mac_address = args.mac_address
-    params.other_info = args.other_info
-    params.serial_number = args.serial_number
-    params.ip_address = args.ip_address
-    params.ip_protocol = args.ip_protocol
-    params.timeout = args.timeout
-    params.file = args.file
-    params.master_board = args.master_board
-    if args.board_id == BoardIds.UNICORN_BOARD:
-        unicorn_devices = retrieve_unicorn_devices()
-
-        args.serial_number = retrieve_unicorn_devices()[2][1]
-    params.serial_number = args.serial_number
-    board_shim = BoardShim(args.board_id, params)
-    try:
-        board_shim.prepare_session()
-        board_shim.start_stream(streamer_params=args.streamer_params)
-        streamer = Streamer(board_shim, params=params, plot=True, save_data=True, window_size=10, update_speed_ms=80)
-    except BaseException:
-        logging.warning('Exception', exc_info=True)
-    finally:
-        logging.info('End')
-        if board_shim.is_prepared():
-            logging.info('Releasing session')
-            try:
-                if streamer.save_data_checkbox.isChecked():
-                    streamer.export_file()
-                board_shim.stop_stream()
-                streamer.lsl_thread.quit()
-            except BaseException:
-                logging.warning('Streaming has already been stopped')
-            board_shim.release_session()
-
-
-if __name__ == '__main__':
-    main()
